+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
+## [1.0.0] - 2021-05-22
+
+### Changed
+
+- Made `LocalDataManager` non `MonoBehaviour`;
+
+### Removed
+
+- Removed unused Editor assembly;
+
 ## [0.0.1] - 2021-02-18
 - Created HephaestusCore Local Data as UnityPackage.